djangopdfmaker/requirements.txt
amqp==5.2.0 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_full_version <= "3.11.2"
billiard==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
brotli==1.1.0 ; platform_python_implementation == "CPython" and python_version >= "3.11" and python_version < "4.0"
brotlicffi==1.1.0.0 ; platform_python_implementation != "CPython" and python_version >= "3.11" and python_version < "4.0"
celery==5.3.5 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cssselect2==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django-filter==23.3 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.7 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
drf-yasg==1.21.7 ; python_version >= "3.11" and python_version < "4.0"
flower==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
fonttools[woff]==4.44.3 ; python_version >= "3.11" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.11" and python_version < "4.0"
humanize==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
kombu==5.3.4 ; python_version >= "3.11" and python_version < "4.0"
markdown==3.5.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pdfkit==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.11" and python_version < "4.0"
prometheus-client==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
prompt-toolkit==3.0.41 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.14.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.1 ; python_version >= "3.11" and python_version < "4.0"
pydyf==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
pyphen==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
redis==5.0.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tinycss2==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.3.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.10 ; python_version >= "3.11" and python_version < "4.0"
weasyprint==60.1 ; python_version >= "3.11" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
zopfli==0.2.3 ; python_version >= "3.11" and python_version < "4.0"
